Refactor utils class into separate files
