+++ index.js
@@ -44,7 +44,7 @@ function wrapParam (p) {
 
 async function run () {
   try {
-    const { repository, version } = await Utils.loadPackage(Utils.cliRoot)
+    const { repository, version } = await Utils.loadPackage(Utils.getCliRoot())
     const repoName = repository.toString().replace('github:', '')
     console.log(`\nRunning ${repoName}@${version}\n`)
