+++ lib/utils/getCliRoot.js
@@ -1,0 +1,5 @@
+import path from 'path'
+
+export default function getCliRoot () {
+  return path.resolve(process.argv[1], '../../at-utils')
+}
