+++ lib/utils/importCore.js
@@ -1,0 +1,6 @@
+import { pathToFileURL } from 'url'
+import path from 'path'
+
+export default function importCore () {
+  return import(pathToFileURL(path.join(root, 'node_modules', 'adapt-authoring-core', 'index.js')))
+}
